requirements.txt
Flask==2.3.3
PyYAML==6.0
openpyxl==3.1.2
gunicorn==22.0.0
